+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 ## Change log
 ----------------------
 
+Version 10
+-------------
+
+CHANGED:
+
+- update to jdk version 21
+
 Version 9.7
 -------------